chore: move setting methods to polyfills
