+++ app/common/renderer/actions/Session.js
@@ -24,11 +24,10 @@ import {
 import {APP_MODE} from '../constants/session-inspector';
 import CloudProviders from '../components/Session/CloudProviders.jsx';
 import i18n from '../i18next';
-import {fs, ipcRenderer, util} from '../polyfills';
+import {fs, ipcRenderer, util, getSetting, setSetting} from '../polyfills';
 import {log} from '../utils/logger';
 import {addVendorPrefixes} from '../utils/other';
 import {quitSession, setSessionDetails} from './Inspector';
-import {getSetting, setSetting} from './settings';
 
 export const NEW_SESSION_REQUESTED = 'NEW_SESSION_REQUESTED';
 export const NEW_SESSION_LOADING = 'NEW_SESSION_LOADING';
